+++ eveuniverse/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "eveuniverse.apps.EveuniverseConfig"
 
-__version__ = "0.7.3"
+__version__ = "0.7.4"
 __title__ = "Eve Universe"